refactor: split cruiser state services by client and server
